ttx: move all components into the libttx
