+++ libs/ttx/terminal.cpp
@@ -1,4 +1,4 @@
-#include "terminal.h"
+#include "ttx/terminal.h"
 
 #include "di/container/algorithm/contains.h"
 #include "di/container/algorithm/rotate.h"
